+++ CHANGES.rst
@@ -19,8 +19,8 @@ Release history
    - Removed
    - Fixed
 
-23.2.24 (unreleased)
-====================
+23.7.30 (July 30, 2023)
+=======================
 
 **Added**